requirements.txt
numpy>=1.24
torch>=2.0
matplotlib>=3.7
plotly>=5.15
pandas>=2.0
pyyaml>=6.0
